// packages/typescript/src/generated/api-types.ts
export interface components {
    schemas: {
// ...
        EvaluateFlag: {
            context?: components["schemas"]["JsonValue"] | null;
        };
// ...
        FlagEvaluationResponse: {
            configuration?: components["schemas"]["JsonValue"] | null;
            context?: components["schemas"]["JsonValue"] | null;
            enabled: boolean;
            key: string;
            scope?: string | null;
            /** Format: int64 */
            timestamp: number;
            variation?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}